+++ R-autofun/check-spsann-arguments.R
@@ -32,7 +32,7 @@ if (any(bb)) {
     res <- c("'candi' must have two named numeric columns: 'x' and 'y'")
   } else {
     # Argument 'iterations'
-    if (!pedometrics::is.numint(iterations) || length(iterations) > 1) {
+    if (!pedometrics::isNumint(iterations) || length(iterations) > 1) {
       res <- c("'iterations' must be an integer value")
     } else {
       # Argument 'acceptance'