+++ NEWS.md
@@ -1,5 +1,6 @@
 # Version 1.0.0.9000
 * Improved and updated repository documentation.
+* Updated dependencies of the package `pedometrics`.
 
 # Version 1.0.0 (2015-07-14)
 * Submission to CRAN.